Change pindefs datatype to uint8_t

+++ pindefs.h
@@ -1,14 +1,14 @@
 #include "Arduino.h"
 
 namespace PinDefs {
-    constexpr int PIN_SLIDE_TOP = A3;
-    constexpr int PIN_SLIDE_BOTTOM = A6;
-    constexpr int PIN_FSR_NECK = A1;
-    constexpr int PIN_FSR_BODY = A0;
-    constexpr int PIN_SWITCH = A2;
-    constexpr int PIN_DIAL = A7;
-    constexpr int PIN_VCA = 3;
-    constexpr int PIN_VCF = 4;
-    constexpr int PIN_VCO = 5;
-    constexpr int PIN_LFO = 6;
+    constexpr uint8_t PIN_SLIDE_TOP = A3;
+    constexpr uint8_t PIN_SLIDE_BOTTOM = A6;
+    constexpr uint8_t PIN_FSR_NECK = A1;
+    constexpr uint8_t PIN_FSR_BODY = A0;
+    constexpr uint8_t PIN_SWITCH = A2;
+    constexpr uint8_t PIN_DIAL = A7;
+    constexpr uint8_t PIN_VCA = 3;
+    constexpr uint8_t PIN_VCF = 4;
+    constexpr uint8_t PIN_VCO = 5;
+    constexpr uint8_t PIN_LFO = 6;
 }
